refactor(desktop): split snapshot tEXt helpers into media_snapshot_png.rs
media.rs crossed the 1000-line desktop file-size limit. Move the
snapshot chunk extract/inject helpers and their tests into a dedicated
module, following the precedent set by media_gif.rs. No behavior
change.
