+++ desktop/src-tauri/src/commands/mod.rs
@@ -25,6 +25,7 @@ mod link_preview;
 pub(crate) mod media;
 mod media_download;
 mod media_gif;
+mod media_snapshot_png;
 mod media_transcode;
 #[cfg(feature = "mesh-llm")]
 pub(crate) mod mesh_llm;
